refactor: 폴더 구조 변경, 테스트 코드 수정
