+++ src/main/java/Application.java
@@ -1,4 +1,4 @@
-import controller.LottoController;
+import lotto.controller.LottoController;
 
 public class Application {
     public static void main(String[] args) {
