+++ src/main/java/lotto/domain/model/LottoNumbers.java
@@ -1,8 +1,8 @@
-package domain.model;
+package lotto.domain.model;
 
 import java.util.List;
 
-import static constant.LottoConstants.*;
+import static lotto.constant.LottoConstants.*;
 
 public class LottoNumbers {
     private final List<Integer> numbers;
